offload (short_)ID creation to util.
